refactor create recommendation to be a component form
